recurring plans

+++ cmd/web/render.go
@@ -26,7 +26,7 @@ var functions = template.FuncMap{
 }
 
 func formatCurrency(n int) string {
-	f := float32(n / 100)
+	f := float32(n) / float32(100)
 	return fmt.Sprintf("%.2f Rs", f)
 }
